+++ src/detectors/near/mod.rs
@@ -1,11 +1,13 @@
 mod frontrunning;
 mod integer_overflow;
 mod missing_deposit_check;
+mod missing_gas_callback;
 mod missing_private;
 mod promise_reentrancy;
 mod self_callback;
 mod signer_vs_predecessor;
 mod storage_staking;
+mod unguarded_storage_unregister;
 mod unhandled_promise;
 mod unsafe_storage_keys;
 
@@ -22,4 +24,8 @@ pub fn register(detectors: &mut Vec<Box<dyn Detector>>) {
     detectors.push(Box::new(frontrunning::FrontrunningDetector));
     detectors.push(Box::new(unsafe_storage_keys::UnsafeStorageKeysDetector));
     detectors.push(Box::new(missing_deposit_check::MissingDepositCheckDetector));
+    detectors.push(Box::new(
+        unguarded_storage_unregister::UnguardedStorageUnregisterDetector,
+    ));
+    detectors.push(Box::new(missing_gas_callback::MissingGasCallbackDetector));
 }